Share the Cargo-env strip, because the sibling test was undoing it
A second adversarial review found that d37ec351 regressed CI, and it was
right. golden_runner.rs ran a bare `Command::new("cargo")` with the Cargo
environment inherited, including the CARGO_MANIFEST_DIR that the previous
commit identifies as the fingerprint-dirtying variable. It runs inside the
same `cargo nextest run -p assay-cli`, concurrently with the wrap tests, so
it re-dirtied the units the wrap tests had just built.

Measured on d37ec351, same command twice with nothing changed between:

  run 1   wrap tests 0.68 / 0.94 / 1.02s   golden_runner 21.9s
  run 2   wrap tests 14.9 / 15.0 / 15.1s   golden_runner 34.1s
  shell `cargo build -p assay-mcp-server` immediately after: 14 crates

That put a nondeterministic multi-crate rebuild inside nextest's
slow-timeout 60s x2 budget, on runners slower than the machine that measured
it, where the failure is a SIGTERM with no message and `retries = 1` buys a
second one. On main these tests asserted a path existed and paid no build
time at all, so this was a regression introduced by this branch, and the CI
pre-warm restored in the previous commit could not have prevented it.

The fix is the rule AGENTS.md already states rather than a bigger timeout:
one rule, one function. "How a test in this crate invokes Cargo" is now
answered once, in tests/common/mod.rs, and both callers use it. golden_runner
also moves to the pinned toolchain via option_env!("CARGO") instead of
whichever cargo is on PATH, and passes --locked, matching the wrap tests.

  run 1   wrap tests 1.27 / 1.39 / 1.47s   golden_runner 10.4s
  run 2   wrap tests 0.71 / 0.81 / 0.94s   golden_runner  0.77s
  run 3   wrap tests 0.73 / 0.84 / 0.90s   golden_runner  0.77s
  shell `cargo build -p assay-mcp-server` immediately after: 0 crates
  whole crate suite: 605 passed, 19.7s -> 2.1s

golden_runner was paying this alternation against shell builds before this
branch existed; sharing the strip fixes that too.

Two smaller findings from the same review:

- target_and_profile canonicalises both paths before comparing. strip_prefix
  is lexical, so a relative CARGO_TARGET_DIR -- which AGENTS.md's per-worktree
  target dir makes likely, not exotic -- or a symlinked one that Cargo reports
  canonicalised would fall through to the profile-only reading and drop the
  triple. That fallback is worse than what it replaced, since it builds a host
  binary where the old code failed loudly, so it is now as narrow as it can be
  made and the residual is stated in the doc comment rather than hidden.

- The two deny fixtures pass against anything spawnable, `/usr/bin/true`
  included; "holds the pipe open" was a sufficient condition described as the
  boundary. Both reviews reached this independently.
